separate the product list into its own file
